Refactor source collector JSON types
Move JSON types of source collectors to a separate module `json_types.py`.
Improve typing.
