+++ components/collector/src/source_collectors/dependency_track/source_up_to_dateness.py
@@ -12,7 +12,8 @@ from collector_utilities.exceptions import CollectorError
 from collector_utilities.type import URL
 from model import Entities, Entity, SourceResponses
 
-from .base import DependencyTrackBase, DependencyTrackProject
+from .base import DependencyTrackBase
+from .json_types import DependencyTrackProject
 
 
 class DependencyTrackSourceUpToDateness(DependencyTrackBase, TimePassedCollector):
