+++ components/collector/src/source_collectors/jenkins/job_runs_within_time_period.py
@@ -5,7 +5,8 @@ from typing import cast
 from collector_utilities.date_time import datetime_from_timestamp, days_ago
 from model import Entities, Entity, SourceMeasurement, SourceResponses
 
-from .base import Build, JenkinsJobs
+from .base import JenkinsJobs
+from .json_types import Build
 
 
 class JenkinsJobRunsWithinTimePeriod(JenkinsJobs):
