+++ components/collector/src/source_collectors/jenkins/source_up_to_dateness.py
@@ -3,7 +3,8 @@
 from collector_utilities.date_time import days_ago
 from model import SourceMeasurement, SourceResponses
 
-from .base import Build, JenkinsJobs
+from .base import JenkinsJobs
+from .json_types import Build
 
 
 class JenkinsSourceUpToDateness(JenkinsJobs):
