+++ components/collector/src/source_collectors/owasp_dependency_check_json/base.py
@@ -1,46 +1,12 @@
 """Base classes for OWASP Dependency-Check JSON collectors."""
 
 from abc import ABC
-from typing import ClassVar, TypedDict
+from typing import ClassVar
 
 from base_collectors import JSONFileSourceCollector
 from collector_utilities.exceptions import JSONAttributeError
 
-
-class Vulnerability(TypedDict):
-    """OWASP Dependency-Check JSON vulnerability."""
-
-    severity: str
-
-
-class Dependency(TypedDict):
-    """OWASP Dependency-Check JSON Dependency."""
-
-    fileName: str
-    filePath: str
-    sha1: str
-    vulnerabilities: list[Vulnerability]
-
-
-class ScanInfo(TypedDict):
-    """OWASP Dependency-Check JSON scan info."""
-
-    engineVersion: str
-
-
-class ProjectInfo(TypedDict):
-    """OWASP Dependency-Check JSON project info."""
-
-    reportDate: str
-
-
-class OWASPDependencyCheckJSON(TypedDict):
-    """OWASP Dependency-Check JSON format."""
-
-    dependencies: list[Dependency]
-    projectInfo: ProjectInfo
-    reportSchema: str
-    scanInfo: ScanInfo
+from .json_types import OWASPDependencyCheckJSON
 
 
 class OWASPDependencyCheckJSONBase(JSONFileSourceCollector, ABC):
